Refactor: Rename the browser_engine header to webview.hpp
There's actually no such thing as a browser_engine, it's merely an alias for the platform-specific implementation that is assigned to it.
